auto_files_tree/requirements.txt
groq==0.3.2
python-dotenv==1.0.0
pyyaml==6.0.1